+++ bindings/dotnet/src/DecentDB.EntityFrameworkCore/Update/Internal/DecentDBModificationCommandBatchFactory.cs
@@ -4,6 +4,13 @@ namespace DecentDB.EntityFrameworkCore.Update.Internal;
 
 internal sealed class DecentDBModificationCommandBatchFactory : IModificationCommandBatchFactory
 {
+    private readonly ModificationCommandBatchFactoryDependencies _dependencies;
+
+    public DecentDBModificationCommandBatchFactory(ModificationCommandBatchFactoryDependencies dependencies)
+    {
+        _dependencies = dependencies;
+    }
+
     public ModificationCommandBatch Create()
-        => new DecentDBModificationCommandBatch();
+        => new DecentDBModificationCommandBatch(_dependencies);
 }